Re-structurize: separate common utils & dictionary
