+++ lattice_tagger/__init__.py
@@ -1,7 +1,3 @@
-from . import beam
-from . import features
-from . import utils
-
 from .tagset import Noun
 from .tagset import Pronoun
 from .tagset import Josa
@@ -14,3 +10,13 @@ from .tagset import Exclamation
 from .tagset import BOS
 from .tagset import EOS
 from .tagset import Unk
+from .utils import installpath
+from .utils import load_rules
+from .utils import write_rules
+from .utils import left_space_tag
+from .utils import get_process_memory
+from .utils import WordMorphemePairs
+
+from . import beam
+from . import dictionary
+from . import features
