+++ lattice_tagger/dictionary/lemmatizer.py
